use k8s cri client as e2e test client
